% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generates a full set of Bayestraits analysis files}
\usage{
generate(mdpath, staging_dir, filepattern = "trait_\%s.dat", commands = c())
}
\arguments{
\item{mdpath}{the path to the directory or metadata JSON file.}

\item{staging_dir}{the path where to write the files.}

\item{filepattern}{the pattern for output files.}
}
\description{
Generates a full set of Bayestraits analysis files
}
