% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_states.R
\name{get_states}
\alias{get_states}
\title{returns a named list of taxon => cognate sets}
\usage{
get_states(cldfobj, word, alphabet = LETTERS, missing_state = "-")
}
\arguments{
\item{cldfobj}{A phlorest object.}

\item{word}{the word to extract states for}

\item{alphabet}{the coding alphabet to use (default=LETTERS)}

\item{missing_state}{the state to use for missing values (default = "-")}
}
\value{
A named list of states
}
\description{
returns a named list of taxon => cognate sets
}
\examples{
phl <- phlorest(system.file("testthat/data/nagaraja_et_al2013", "cldf-metadata.json", package = "rcldf"))
get_states(phl, 'NAME')
}
