% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treefile.R
\name{get_treefile}
\alias{get_treefile}
\title{Returns the file path to the phylogenies in the dataset}
\usage{
get_treefile(cldfobj, type = "posterior")
}
\arguments{
\item{cldfobj}{A rcldf::cldf/phlorest object.}

\item{type}{A string - "best", "summary", "posterior".}
}
\value{
A file path
}
\description{
Returns the file path to the phylogenies in the dataset
}
\examples{
mdpath <- system.file("testthat/data", "cldf-metadata.json", package = "wordrates")
phl <- read.phlorest(mdpath)
get_treefile(phl, 'posterior')
'testthat/data/posterior.trees'
get_treefile(phl, 'summary')
'testthat/data/summary.trees'
}
