% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.constant}
\alias{is.constant}
\title{Checks if a state object is constant}
\usage{
is.constant(states)
}
\arguments{
\item{states}{A list of states}
}
\value{
Boolean TRUE/FALSE
}
\description{
Checks if a state object is constant
}
\examples{
is.constant(list('L1' = c("A"), 'L2' = c("A"))) == TRUE
is.constant(list('L1' = c("A"), 'L2' = c("-"))) == TRUE
is.constant(list('L1' = c("A"), 'L2' = c("B"))) == FALSE
}
