% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.phlorest.R
\name{phlorest}
\alias{phlorest}
\title{Loads a Phlorest Dataset.}
\usage{
phlorest(mdpath, datafile = "data.nex")
}
\arguments{
\item{mdpath}{the path to the directory or metadata JSON file.}

\item{datafile}{the expected datafile name (default="data.nex")}
}
\value{
A \code{phlorest} object
}
\description{
Loads a Phlorest Dataset.
}
\examples{
mdpath <- system.file("testthat/data", "cldf-metadata.json", package = "wordrates")
phl <- phlorest(mdpath)
}
