% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.phlorest.R
\name{read.phlorest}
\alias{read.phlorest}
\title{Loads a Phlorest Dataset.}
\usage{
read.phlorest(metadata, datafile = "data.nex")
}
\arguments{
\item{mdpath}{the path to the directory or metadata JSON file.}
}
\value{
A \code{phlorest} object
}
\description{
Loads a Phlorest Dataset.
}
\examples{
phl <- cldf(system.file("testthat/data/nagaraja_et_al2013", "cldf-metadata.json", package = "rcldf"))
}
