% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordrates-package.R
\docType{package}
\name{wordrates-package}
\alias{wordrates}
\alias{wordrates-package}
\title{wordrates: Calculate WordRates}
\description{
WordRates Project Code. Not really for public consumption.
}
\author{
\strong{Maintainer}: Simon Greenhill \email{simon@simon.net.nz} (\href{https://orcid.org/0000-0001-7832-6156}{ORCID})

}
\keyword{internal}
