% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.bayestraits.R
\name{write.bayestraits}
\alias{write.bayestraits}
\title{writes a file in Bayestraits form}
\usage{
write.bayestraits(states, filename)
}
\arguments{
\item{states}{a states object from get_states}

\item{filename}{filename to write to.}
}
\description{
writes a file in Bayestraits form
}
