% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generates a full set of Bayestraits analysis files}
\usage{
generate(metadata, staging_dir, filepattern = "trait_\%s.dat")
}
\arguments{
\item{staging_dir}{the path where to write the files}

\item{mdpath}{the path to the directory or metadata JSON file.}
}
\description{
Generates a full set of Bayestraits analysis files
}
