% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_states.R
\name{get_states}
\alias{get_states}
\title{returns a named list of taxon => cognate sets}
\usage{
get_states(cldfobj, word, alphabet = ALPHABET, missing_state = "-")
}
\arguments{
\item{cldfobj}{A phlorest object.}

\item{word}{the word to extract states for}

\item{alphabet}{the coding alphabet to use (default=c(LETTERS, 1:9, 0))}

\item{missing_state}{the state to use for missing values (default = "-")}
}
\value{
A named list of states
}
\description{
returns a named list of taxon => cognate sets
}
\examples{
mdpath <- system.file("testthat/data", "cldf-metadata.json", package = "wordrates")
get_states(read.phlorest(mdpath), 'NAME')
}
