% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treefile.R
\name{get_treefile}
\alias{get_treefile}
\title{Returns the file path to the phylogenies in the dataset}
\usage{
get_treefile(cldfobj, type = "posterior")
}
\arguments{
\item{cldfobj}{A rcldf::cldf/phlorest object.}

\item{type}{A string - "best", "summary", "posterior".}
}
\value{
A file path
}
\description{
Returns the file path to the phylogenies in the dataset
}
\examples{
cldfobj <- cldf(system.file("testthat/data/nagaraja_et_al2013", "cldf-metadata.json", package = "rcldf"))
get_treefile(cldfobj, 'posterior')
'testthat/data/nagaraja_et_al2013/posterior.trees'
get_treefile(cldfobj, 'summary')
'testthat/data/nagaraja_et_al2013/summary.trees'
}
