% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{get_words}
\alias{get_words}
\title{Extracts a list of words from the phlorest dataset}
\usage{
get_words(cldfobj)
}
\arguments{
\item{cldfobj}{A rcldf::cldf/phlorest object.}
}
\value{
A list of words
}
\description{
Extracts a list of words from the phlorest dataset
}
\examples{
cldfobj <- cldf(system.file("testthat/data/nagaraja_et_al2013", "cldf-metadata.json", package = "rcldf"))
get_words(cldfobj)
}
